cims-python-develop/requirements.txt
apiclient==1.0.4
asn1crypto==0.24.0
beautifulsoup4==4.7.1
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.2
django-cors-headers==3.0.2
django-tinymce4-lite==1.7.5
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
get==2019.4.13
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gspread==3.1.0
httplib2==0.13.0
idna==2.8
jsmin==2.2.2
lxml==4.3.4
oauth2client==4.1.3
post==2019.4.13
public==2019.4.13
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
